+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Do not use public SSH keys coming from the CRs.
+
+### Fixed
+
+- Try to send only one request to VMSS Azure API from `nodepool` handler.
+
 ## [5.0.0-alpha2] - 2020-10-14
 
 ### Fixed